TOTSE - Temple of the Screaming Electron
https://newtotse.com/oldtotse/en/_faq/totse_faq.html

A lot of people have some weird idea that this web site is a Bad Place, a place for hackers, software pirates, and anarchists. The reason that they think this is that there are informational text files on here about hacking, piracy, and anarchy.

However, there are also text files on here that discuss politics; democratic, right wing, left wing, libertarian, communist, and everything in between, but this is not a political web site.

There are files on here that discuss Jesus Christ, Muhammed, Buddha, Crowley, John Smith, and "Bob", but this is not a religious web site.

There are files full of short stories, science fiction, humorous articles, and great works of literature, but this is not a literary web site.

There are files with information on rocketry, radio broadcasting, chemistry, electronics, genetics, and computers, but this is not a technical web site.

This web site is about INFORMATION. All sorts and all viewpoints. Some of the information you will agree with, some you will find shocking, and some you will probably disagree with violently. That is the whole point. In this society we go to schools where there is one right answer: The Teacher's. There is one acceptable version of events: The Television's. There is only one acceptable occupation: The pursuit of money. There is only one political choice to make: The Status Quo.

On this web site you are expected to make decisions all by yourself. You get to decide who and what to agree with, and why. You get to hear new viewpoints that you may have never heard before. On this web site people exist without age, without skin color, without gender, without clothes, without nationality, without any of the visual cues we usually use to discredit or ignore people who are unlike ourselves. All of these things are stripped away and the ideas themselves are laid bare.

You will change. You will transform. You will learn. You will disagree.

You will enjoy it.